+++ SuspensionAnalysis/src/SuspensionAnalysis.Core/Mapper/MappingResolver.cs
@@ -25,14 +25,14 @@ namespace SuspensionAnalysis.Core.Mapper
             return new SuspensionSystem
             {
                 ShockAbsorber = ShockAbsorber.Create(request.ShockAbsorber),
-                SuspensionAArmLower = SuspensionAArm.Create(request.SuspensionAArmLower),
-                SuspensionAArmUpper = SuspensionAArm.Create(request.SuspensionAArmUpper),
+                LowerWishbone = SuspensionWishbone.Create(request.LowerWishbone),
+                UpperWishbone = SuspensionWishbone.Create(request.UpperWishbone),
                 TieRod = TieRod.Create(request.TieRod)
             };
         }
 
         /// <summary>
-        /// This method craetes a <see cref="SuspensionAArm{TProfile}"/> based on <see cref="RunAnalysisRequest{TProfile}"/> and <see cref="CalculateReactionsResponseData"/>.
+        /// This method craetes a <see cref="SuspensionWishbone{TProfile}"/> based on <see cref="RunAnalysisRequest{TProfile}"/> and <see cref="CalculateReactionsResponseData"/>.
         /// </summary>
         /// <typeparam name="TProfile"></typeparam>
         /// <param name="runAnalysisRequest"></param>
@@ -49,8 +49,8 @@ namespace SuspensionAnalysis.Core.Mapper
             return new SuspensionSystem<TProfile>
             {
                 ShockAbsorber = ShockAbsorber.Create(runAnalysisRequest.ShockAbsorber, calculateReactionsResponseData.ShockAbsorberReaction.AbsolutValue),
-                SuspensionAArmLower = SuspensionAArm<TProfile>.Create(runAnalysisRequest.SuspensionAArmLower, runAnalysisRequest.Material, calculateReactionsResponseData.AArmLowerReaction1.AbsolutValue, calculateReactionsResponseData.AArmLowerReaction2.AbsolutValue),
-                SuspensionAArmUpper = SuspensionAArm<TProfile>.Create(runAnalysisRequest.SuspensionAArmUpper, runAnalysisRequest.Material, calculateReactionsResponseData.AArmUpperReaction1.AbsolutValue, calculateReactionsResponseData.AArmUpperReaction2.AbsolutValue),
+                LowerWishbone = SuspensionWishbone<TProfile>.Create(runAnalysisRequest.LowerWishbone, runAnalysisRequest.Material, calculateReactionsResponseData.LowerWishboneReaction1.AbsolutValue, calculateReactionsResponseData.LowerWishboneReaction2.AbsolutValue),
+                UpperWishbone = SuspensionWishbone<TProfile>.Create(runAnalysisRequest.UpperWishbone, runAnalysisRequest.Material, calculateReactionsResponseData.UpperWishboneReaction1.AbsolutValue, calculateReactionsResponseData.UpperWishboneReaction2.AbsolutValue),
                 TieRod = TieRod<TProfile>.Create(runAnalysisRequest.TieRod, runAnalysisRequest.Material, calculateReactionsResponseData.TieRodReaction.AbsolutValue)
             };
         }