+++ SuspensionAnalysis/src/SuspensionAnalysis.DataContracts/Models/Analysis/SuspensionWishboneAnalysisResult.cs
@@ -3,30 +3,30 @@
 namespace SuspensionAnalysis.DataContracts.Models.Analysis
 {
     /// <summary>
-    /// It contains the necessary informations to execute the analysis to suspension A-arm.
+    /// It contains the necessary informations to execute the analysis to suspension wihsbone.
     /// </summary>
-    public class SuspensionAArmAnalysisResult
+    public class SuspensionWishboneAnalysisResult
     {
         /// <summary>
-        /// The critical buckling force to one segment of suspension A-arm.
+        /// The critical buckling force to one segment of suspension wishbone.
         /// Unit: N (Newton).
         /// </summary>
         public double CriticalBucklingForce1 { get; set; }
 
         /// <summary>
-        /// The critical buckling force to another segment of suspension A-arm.
+        /// The critical buckling force to another segment of suspension wishbone.
         /// Unit: N (Newton).
         /// </summary>
         public double CriticalBucklingForce2 { get; set; }
 
         /// <summary>
-        /// The applied force to one segment of suspension A-arm.
+        /// The applied force to one segment of suspension wishbone.
         /// Unit: N (Newton).
         /// </summary>
         public double AppliedForce1 { get; set; }
 
         /// <summary>
-        /// The applied force to another segment of suspension A-arm.
+        /// The applied force to another segment of suspension wishbone.
         /// Unit: N (Newton).
         /// </summary>
         public double AppliedForce2 { get; set; }