+++ SuspensionAnalysis/src/SuspensionAnalysis.DataContracts/Models/Enums/BearingType.cs
@@ -1,0 +1,10 @@
+﻿namespace SuspensionAnalysis.DataContracts.Models.Enums
+{
+    /// <summary>
+    /// It represents the bearing used in suspension system. 
+    /// </summary>
+    public enum BearingType
+    {
+        Bearing1 = 1 
+    }
+}