+++ SuspensionAnalysis/test/SuspensionAnalysis.UnitTest/Helper/RunAnalysisHelper.cs
@@ -24,27 +24,27 @@ namespace SuspensionAnalysis.UnitTest.Helper.DataContracts
                     FasteningPoint = "-0.005,0.645,0.180",
                     PivotPoint = "-0.005,0.485,0.430"
                 },
-                SuspensionAArmLower = new SuspensionAArm<CircularProfile>
+                LowerWishbone = new SuspensionWishbone<CircularProfile>
                 {
                     Profile = new CircularProfile
                     {
                         Diameter = 25.4e-3,
                         Thickness = 0.9e-3
                     },
-                    KnucklePoint = "-0.012,0.685,0.150",
-                    PivotPoint1 = "0.250,0.350,0.150",
-                    PivotPoint2 = "-0.100,0.350,0.130"
+                    WishboneOuterBallJoint = "-0.012,0.685,0.150",
+                    WishboneFrontPivot = "0.250,0.350,0.150",
+                    WishboneRearPivot = "-0.100,0.350,0.130"
                 },
-                SuspensionAArmUpper = new SuspensionAArm<CircularProfile>
+                UpperWishbone = new SuspensionWishbone<CircularProfile>
                 {
                     Profile = new CircularProfile
                     {
                         Diameter = 25.4e-3,
                         Thickness = 0.9e-3
                     },
-                    KnucklePoint = "0.012,0.660,0.410",
-                    PivotPoint1 = "0.200,0.450,0.362",
-                    PivotPoint2 = "-0.080,0.450,0.362"
+                    WishboneOuterBallJoint = "0.012,0.660,0.410",
+                    WishboneFrontPivot = "0.200,0.450,0.362",
+                    WishboneRearPivot = "-0.080,0.450,0.362"
                 },
                 TieRod = new TieRod<CircularProfile>
                 {
@@ -73,7 +73,7 @@ namespace SuspensionAnalysis.UnitTest.Helper.DataContracts
                     FasteningPoint = "-0.005,0.645,0.180",
                     PivotPoint = "-0.005,0.485,0.430"
                 },
-                SuspensionAArmLower = new SuspensionAArm<RectangularProfile>
+                LowerWishbone = new SuspensionWishbone<RectangularProfile>
                 {
                     Profile = new RectangularProfile
                     {
@@ -81,11 +81,11 @@ namespace SuspensionAnalysis.UnitTest.Helper.DataContracts
                         Width = 10e-3,
                         Thickness = 1e-3
                     },
-                    KnucklePoint = "-0.012,0.685,0.150",
-                    PivotPoint1 = "0.250,0.350,0.150",
-                    PivotPoint2 = "-0.100,0.350,0.130"
+                    WishboneOuterBallJoint = "-0.012,0.685,0.150",
+                    WishboneFrontPivot = "0.250,0.350,0.150",
+                    WishboneRearPivot = "-0.100,0.350,0.130"
                 },
-                SuspensionAArmUpper = new SuspensionAArm<RectangularProfile>
+                UpperWishbone = new SuspensionWishbone<RectangularProfile>
                 {
                     Profile = new RectangularProfile
                     {
@@ -93,9 +93,9 @@ namespace SuspensionAnalysis.UnitTest.Helper.DataContracts
                         Width = 10e-3,
                         Thickness = 1e-3
                     },
-                    KnucklePoint = "0.012,0.660,0.410",
-                    PivotPoint1 = "0.200,0.450,0.362",
-                    PivotPoint2 = "-0.080,0.450,0.362"
+                    WishboneOuterBallJoint = "0.012,0.660,0.410",
+                    WishboneFrontPivot = "0.200,0.450,0.362",
+                    WishboneRearPivot = "-0.080,0.450,0.362"
                 },
                 TieRod = new TieRod<RectangularProfile>
                 {