+++ SuspensionAnalysis/etc/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+### Added
+ - Operation CalculateSteeringKnuckleReactions. 
+### Changed 
+ - Renamed A-Arm to Wishbone in all application.  
+ - Renamed pivotPoint1 to wishboneFrontPivot in all application.
+ - Renamed pivotPoint2 to wishboneRearPivot in all application.
+ - Renamed knucklePoint to wishboneOuterBallJoint in all application. 
+
+
 ## [1.0.0] - 2021-04-20
 ### Added
  - First version of the program.